requirements.txt
flask==1.1.2
flask-caching-1.10.1